Home.js refactoring
